+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 1.4.0.0 - 2021-11-??
 * Plugin renamed to **Astro-Physics Tools** - same great taste, all-new flavor. But, really, the original name was a mouthful.
 * Renamed the **Create APPM Model** instruction to **Create All-Sky Model**
-* Initial beta of the new **Create Dec Arc Model** instruction 
+* Initial beta of the new **Create Dec Arc Model** instruction
+* Instructions now check to make sure a camera is connected in NINA (for APPM's use) and raise a validation warning if one is not
 * Minimum supported NINA version is now 2.0 Beta 4
 
 ## 1.3.5.0 - 2021-10-27